# docs/swagger.yaml
definitions:
...
  github_com_bsonger_devflow-config-service_pkg_model.JobStatus:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    - RollingBack
    - RolledBack
    - Syncing
    - SyncFailed
    type: string
    x-enum-varnames:
    - JobPending
    - JobRunning
    - JobSucceeded
    - JobFailed
    - JobRollingBack
    - JobRolledBack
    - JobSyncing
    - JobSyncFailed
...
  github_com_bsonger_devflow-config-service_pkg_model.ManifestStatus:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    type: string
    x-enum-varnames:
    - ManifestPending
    - ManifestRunning
    - ManifestSucceeded
    - ManifestFailed
...
  github_com_bsonger_devflow-config-service_pkg_model.StepStatus:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    type: string
    x-enum-varnames:
    - StepPending
    - StepRunning
    - StepSucceeded
    - StepFailed
...
  pkg_api.VerifyBuildStatusRequest:
    properties:
      external_ref:
        type: string
      intent_id:
        type: string
      manifest_id:
        type: string
      message:
        type: string
      pipeline_id:
        type: string
      status:
        $ref: '#/definitions/github_com_bsonger_devflow-config-service_pkg_model.ManifestStatus'
    required:
    - manifest_id
    - status
    type: object
  pkg_api.VerifyBuildStepRequest:
    properties:
      end_time:
        type: string
      manifest_id:
        type: string
      message:
        type: string
      pipeline_id:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_bsonger_devflow-config-service_pkg_model.StepStatus'
      task_name:
        type: string
      task_run:
        type: string
    required:
    - manifest_id
    - status
    - task_name
    type: object
  pkg_api.VerifyReleaseStatusRequest:
    properties:
      external_ref:
        type: string
      intent_id:
        type: string
      job_id:
        type: string
      message:
        type: string
      status:
        $ref: '#/definitions/github_com_bsonger_devflow-config-service_pkg_model.JobStatus'
    required:
    - job_id
    - status
    type: object
  pkg_api.VerifyReleaseStepRequest:
    properties:
      end_time:
        type: string
      job_id:
        type: string
      message:
        type: string
      progress:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_bsonger_devflow-config-service_pkg_model.StepStatus'
      step_name:
        type: string
    required:
    - job_id
    - status
    - step_name
    type: object
info:
  contact: {}
paths:
...
  /api/v1/verify/argo/events:
    post:
      consumes:
      - application/json
      description: 由 Argo 或外部发布观察器回写 Job 级状态
      parameters:
      - description: Release Status Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_api.VerifyReleaseStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 回写发布状态
      tags:
      - Verify
...
  /api/v1/verify/release/steps:
    post:
      consumes:
      - application/json
      description: 由 Argo Application / Deployment / Rollout 观察器回写 Job steps
      parameters:
      - description: Release Step Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_api.VerifyReleaseStepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 回写发布步骤
      tags:
      - Verify
  /api/v1/verify/tekton/events:
    post:
      consumes:
      - application/json
      description: 由 Tekton 或外部构建观察器回写 Manifest 级状态
      parameters:
      - description: Build Status Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_api.VerifyBuildStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 回写构建状态
      tags:
      - Verify
  /api/v1/verify/tekton/steps:
    post:
      consumes:
      - application/json
      description: 由 Tekton TaskRun 观察器回写 Manifest steps
      parameters:
      - description: Build Step Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_api.VerifyBuildStepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 回写构建步骤
      tags:
      - Verify
swagger: "2.0"